do not modify parameters unless explicitly set
The contrast, saturation, brightness, sharpness and focus parameters
were recently added to usb_cam. This caused a regression
(sigproc/robotic_surgery#17) whereby the default settings for a webcam
are overridden in all cases by the hard-coded defaults in usb_cam.

In the absence of a know good set of "default" values, leave the
parameters unset unless the user has explicitly set them in the launch
file.