Add ROS parameter for tuning MJPEG video streaming
MJPEG video frame format can have multiple pixel formats and we need to
provide a way to allow the user to set up the appropriate one for
establishing image buffers in the right way. The most common pixel
formats supported by ffmpeg codec are YUV420P, a 12 bits-per-pixel
encoding, and YUV422P, a 16 bits-per-pixel encoding. We are adding a
new integer ROS parameter to control the selection of the both most
used encodings for MJPEG: 'bits_per_pixel'. Only two values are valid as
today: 12 and 16; the former is the default.