remove AV_PIX_FMT_XVMC
This no longer exists in ffmpeg trunk. I see that is exists in older
versions, such as 2.7. However, here it is not documented.

There are simmilarly named and now document formats such as AV_PIX_FMT_XV30
and AV_PIX_FMT_XV60, which should probably be supported instread, but I
will not add these in this commit.